Change HomePage inside post entity
