+++ src/router/index.tsx
@@ -5,7 +5,7 @@ import {
   Route,
 } from "react-router-dom";
 import App from "../components/App/App";
-import HomePage from "../pages/HomePage/HomePage";
+import HomePage from "../post/pages/HomePage/HomePage";
 import NotFoundPage from "../pages/NotFoundPage/NotFoundPage";
 
 const router = createBrowserRouter(
