+++ src/styles/styles.css
@@ -43,6 +43,6 @@ p {
   margin: 0;
 }
 
-.page-titles {
+.page-title {
   margin-bottom: 30px;
 }
